analysis-service/requirements.txt
# Core ML / Data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.11.0

# NLP / Text Processing
nltk>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Explainability
shap>=0.43.0
lime>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wordcloud>=1.9.0

# Demo UI
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0
python-dotenv>=1.0.0

# Email Parsing
chardet>=5.2.0

# Optional: DistilBERT (install at Deep Learning phase)
# torch>=2.0.0
# transformers>=4.35.0
# datasets>=2.14.0
